feat: CommentCard 완료 및 Profile class 위치 변경

+++ client/app/src/main/java/com/org/egglog/client/ui/molecules/profileButtonList/ProfileButtonList.kt
@@ -18,7 +18,7 @@ import com.org.egglog.client.ui.atoms.buttons.ProfileButton
 import com.org.egglog.client.utils.widthPercent
 
 @Composable
-fun ProfileButtonList(userInfoList: ArrayList<UserInfo>, selectedList: SnapshotStateList<Int>, myUserId: Int) {
+fun ProfileButtonList(userInfoList: ArrayList<UserInfo>, selectedList: SnapshotStateList<Long>, myUserId: Long) {
     Row(
         modifier = Modifier
             .fillMaxWidth()
